LU-5417 lustre: fix comparison between signed and unsigned
Cleanup in general headers.
* use size_t in cfs_size_round*()
* make unsigned index and len in lustre_cfg_*() and object_update_*()
* make iteration variable the same type as comparing value
* make unsigned pages counters
* fix printf formats

Reviewed-on: http://review.whamcloud.com/11327
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/ptlrpc/pack_generic.c
@@ -1929,7 +1929,7 @@ EXPORT_SYMBOL(lustre_swab_mgs_target_info);
 
 void lustre_swab_mgs_nidtbl_entry(struct mgs_nidtbl_entry *entry)
 {
-        int i;
+	__u8 i;
 
         __swab64s(&entry->mne_version);
         __swab32s(&entry->mne_instance);
@@ -2031,7 +2031,7 @@ void lustre_swab_fiemap_extent(struct ll_fiemap_extent *fm_extent)
 
 void lustre_swab_fiemap(struct ll_user_fiemap *fiemap)
 {
-        int i;
+	__u32 i;
 
         __swab64s(&fiemap->fm_start);
         __swab64s(&fiemap->fm_length);
@@ -2560,7 +2560,7 @@ EXPORT_SYMBOL(lustre_swab_update_buf);
 
 void lustre_swab_update_reply_buf(struct update_reply *ur)
 {
-	int i;
+	size_t i;
 
 	__swab32s(&ur->ur_version);
 	__swab32s(&ur->ur_count);
